app1/pyproject.toml
[tool.poetry]
name = "app1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
opentelemetry-distro = "^0.44b0"
opentelemetry-exporter-otlp = "^1.23.0"
sqlalchemy = "^2.0.27"
psycopg = {extras = ["binary"], version = "^3.1.18"}
opentelemetry-instrumentation-sqlalchemy = "^0.44b0"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"